backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy>=2.0.35
psycopg2-binary==2.9.10
python-dotenv==1.0.0
pdfplumber==0.11.0
Pillow>=10.0.0
requests>=2.31.0

